scraper/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
coverage==4.5.4
cssselect==1.1.0
dnspython==1.16.0
feedparser==5.2.1
idna==2.8
lxml==4.4.1
multidict==4.5.2
nose2==0.9.1
pymongo==3.8.0
pyquery==1.4.0
PyYAML==5.1.2
six==1.12.0
yarl==1.3.0
